path: page component moved to the PageComp folder
